Create find_by_order() .CPP
The find_by_order() is a built-in function of Ordered Set which is a Policy-Based Data Structure in C++. Policy-based data structures are not part of the C++ Standard Template Library but the g++ compiler supports them.

Ordered Set is a policy-based data structure in g++ that maintains unique elements in sorted order. It performs all the operations as performed by Set in STL in O(logN) complexity. 
In addition to that, the following two operations are also performed in O(logN) complexity:

order_of_key (K): Number of items strictly smaller than K.
find_by_order(k): Kth element in a Set (counting from zero).
The find_by_order() function accepts a key, say K, as an argument and returns the iterator to the Kth largest element in the Set.