Update quadratic.cpp
The code had a logical error in the calculation of the roots of a quadratic equation. Specifically, the formulas used to calculate the roots are incorrect due to missing parentheses. So I have corrected that.

This commit is meant for the purpose of hacktoberfest, please do accept the PR  with hacktoberfest tag on it.