Create QuickSort
This code defines a quickSort function that takes an array as input and recursively divides it into smaller subarrays by choosing a pivot element and rearranging the elements so that all elements less than the pivot are on the left and all elements greater than the pivot are on the right. The process continues until the base case is reached (arrays of size 0 or 1). Finally, it concatenates the sorted subarrays to produce the sorted result.